調整

+++ setup.sh
@@ -1,4 +1,7 @@
 #!/bin/sh
-ln -sd ~/dotfiles/.vim ~/.vim
+if [ ! -L ~/.vim ]; then
+	ln -sd ~/dotfiles/.vim ~/.vim
+fi
 ln -sf ~/dotfiles/.vimrc ~/.vimrc
 ln -sf ~/dotfiles/.bashrc ~/.bashrc
+ln -sf ~/dotfiles/.gitconfig ~/.gitconfig
